onMenu 위치 이동 및 onMenu 폴더 삭제
